Updated Vulkan test #1 - most of features implemented. To be optimized.
Most of test #1 is implemented in Vulkan. There are some details to be
added, fixed and/or optimized, as current implementation is more like
proof-of-conecpt then optimized real-life version.

OpenGL test #1 is almost finished, needs some polishing with measurements
and should be ready to use soon. Vulkan's version still needs some work.